requirements.txt
# データ取得
yfinance>=1.2.0
requests>=2.31.0
openpyxl>=3.1.2          # JPX の Excel ファイル読み込み

# データ処理
pandas>=2.2.0
numpy>=1.26.0

# テクニカル指標
ta>=0.11.0

# 機械学習（フェーズ1）
lightgbm>=4.3.0
scikit-learn>=1.4.0
optuna>=4.0.0               # 多目的ハイパーパラメータ最適化

# 深層学習（フェーズ2: TFT）
torch>=2.2.0
pytorch-lightning>=2.2.0
pytorch-forecasting>=1.0.0

# 並列処理
joblib>=1.3.0

# スケジューリング
APScheduler>=3.10.0

# 可視化
matplotlib>=3.8.0
plotly>=5.20.0

# ユーティリティ
python-dotenv>=1.0.0
jpbizday>=0.0.1           # 日本の営業日カレンダー
tqdm>=4.66.0              # 進捗表示
